refactor(ContextMenuView): rename ContextMenuContent to ContextMenuView
