--- packages/code-editor/src/ContextMenu/ContextMenuContent/index.ts
@@ -1,6 +1,0 @@
-export { ContextMenuContent } from './ContextMenuContent';
-export type {
-  ContextMenuContentProps,
-  MenuItem,
-  MenuState,
-} from './ContextMenuContent.types';
