+++ packages/code-editor/src/ContextMenu/ContextMenuView/index.ts
@@ -1,0 +1,6 @@
+export { ContextMenuView } from './ContextMenuView';
+export type {
+  ContextMenuContentProps,
+  MenuItem,
+  MenuState,
+} from './ContextMenuView.types';
